Header ready (exception: animations not written yet)
The block metric has been rewritten with more readable units, the block hierarchy has been rebuilt in "index.html". In this version, the <a> tag contains a <figure> tag instead of a <div> tag. Also changed the font for the buttons in the menu (M PLUS Rounded 1c -> Roboto).

The issue that should be resolved is how to achieve the same "header-banner" position via "padding" as it currently is (now achieved by decreasing the block's width as a percentage and padding: 0). If "padding" is applied, the "banner-brand-link" and "banner-feedback-link" blocks are glued together even though "justify-content: space-between" is applied. Perhaps the problem is that the "header-banner" block is now a child block. The clumsy solution is marked as "/* marked place */" (description of ".header-banner" in "style.css".