All pages of the site are made
The layout of the remaining pages has been written in addition to the main one. Fixed a bug with going beyond the borders of the block frame with editors on the main page. The wave background now doesn't change size when scaling on the main page and in the archive